Move plotting outside stages
