+++ snn_hfo_ieeg/plotting/plot_patient.py
